data_viz/sw_boston_peer_cities_dashboard/requirements.txt
dash
geopandas
plotly-express
gunicorn
pyarrow
